MLI_Architect/requirements.txt
aiogram>=3.0             # Telegram Bot API
pydantic>=2.0            # Валидация данных и настроек
python-dotenv            # Чтение .env
aiohttp                  # Асинхронные запросы (для парсера)
beautifulsoup4           # Очистка HTML (для парсера)
groq                     # Клиент для Llama-3.3
aiosqlite                # Асинхронная база данных
loguru                   # Красивое логирование